Remove useless milestone_timestamp as u32 casts

+++ bee-ledger/src/workers/consensus/white_flag.rs
@@ -90,11 +90,7 @@ fn apply_regular_essence<B: StorageBackend>(
             output_id,
             (
                 created_output,
-                ConsumedOutput::new(
-                    *transaction_id,
-                    metadata.milestone_index,
-                    metadata.milestone_timestamp as u32,
-                ),
+                ConsumedOutput::new(*transaction_id, metadata.milestone_index, metadata.milestone_timestamp),
             ),
         );
     }
@@ -105,7 +101,7 @@ fn apply_regular_essence<B: StorageBackend>(
             CreatedOutput::new(
                 *message_id,
                 metadata.milestone_index,
-                metadata.milestone_timestamp as u32,
+                metadata.milestone_timestamp,
                 output.clone(),
             ),
         );
